rename: 알림 관련 패키지 변경
