+++ presentation/src/main/java/com/hihihihi/gureumpage/notification/common/Channels.kt
@@ -1,4 +1,4 @@
-package com.hihihihi.gureumpage.notification
+package com.hihihihi.gureumpage.notification.common
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
